+++ farmerconnect/lib/farmmer/feed/feedRequests.dart
@@ -29,7 +29,7 @@ class feedRequests extends StatelessWidget {
           return feedRequestsModel(
               TalepID: map["TalepID"],
               YemAdi: map['YemAdı'],
-              Miktar: map["Miktar"],
+              Miktar: map["Miktar"].toDouble(),
               Durum: map["Durum"],
               IstekTarihi: map["IstekTarihi"],
               TeslimTarihi: map["TeslimTarihi"]);
